# Expand `if_any()` and `if_all()` expressions