--- a/src/init.cpp
+++ b/src/init.cpp
@@ -113,9 +113,6 @@ static const R_CallMethodDef CallEntries[] = {
   {"dplyr_mask_remove", (DL_FUNC)& dplyr_mask_remove, 2},
   {"dplyr_mask_add", (DL_FUNC)& dplyr_mask_add, 3},
 
-  {"dplyr_reduce_lgl_or", (DL_FUNC)& dplyr_reduce_lgl_or, 2},
-  {"dplyr_reduce_lgl_and", (DL_FUNC)& dplyr_reduce_lgl_and, 2},
-
   {"dplyr_lazy_vec_chop_impl", (DL_FUNC)& dplyr_lazy_vec_chop, 2},
   {"dplyr_data_masks_setup", (DL_FUNC)& dplyr_data_masks_setup, 3},
   {"env_resolved", (DL_FUNC)& env_resolved, 2},